move maps to own folder
